requirements.txt
asgiref==3.2.7
astroid==2.7.3
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.5
django-cors-headers==3.7.0
django-environ==0.5.0
django-filter==2.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.7.1
drf-yasg==1.20.0
flake8==3.9.2
flake8-isort==4.0.0
flake8-plugin-utils==1.3.2
flake8-return==1.1.3
gunicorn==20.1.0
idna==2.9
importlib-metadata==1.6.0
inflection==0.5.1
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
platformdirs==2.3.0
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.8.1
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
pylint==2.10.2
pyparsing==2.4.7
pytest==5.4.1
pytest-django==3.9.0
python-dotenv==0.17.1
pytz==2019.3
requests==2.26.0
ruamel.yaml==0.17.9
ruamel.yaml.clib==0.2.2
six==1.14.0
sqlparse==0.3.1
testfixtures==6.18.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.1.9
wrapt==1.12.1
zipp==3.1.0
